refactor(geometry): split router/transforms.rs into 5 placement modules
transforms.rs (1210 LOC) -> transforms/{mod (flags + IfcLocalPlacement
hierarchy dispatch), parsers (leaf IFC primitive placement parsers), linear
(#859 linear-placement path), grid (#883 grid path), mesh_world (f32
local-frame mesh transforms)}. All files under ~400.

Pure code motion: the inherent impl GeometryRouter is split across files as
multiple impl blocks. Items called by router siblings (parse_*, transform_mesh_*)
become pub(crate); intra-transforms cross-file calls pub(super); router-level
refs adjusted super:: -> super::super::. router/mod.rs untouched; processing.rs's
super::transforms::{instancing_enabled,mat4_to_row_major} resolves unchanged.
build/clippy -D warnings clean; transform/grid/linear/placement tests green.
